build-installers: include /etc/msystem.d when available
In 831610377 (make-file-list: unbreak `nuget/release.sh`, 2025-06-03), I
introduced a check for that directory; Contingent on this, its contents
are included in the file list that is used to build the installers.

However, we need to imitate 2141a50b2 (make-file-list: do not expect
/etc/bash.bash_logout to exist, 2025-02-26). That commit guarantees that
the `bash_logout` file is checked out if it exists in the HEAD revision
(so that the check for its presence in `make-file-list.sh` will not fail
by mistake). Likewise, we need to ensure that `/etc/msystem.d/` is
checked out (if it exists in HEAD).

This bug prevented the Windows/ARM64 variant of Git for Windows'
installer from containing the `/etc/msystem.d/` directory, which
results in the annoying error message "bash: /etc/msystem.d/MSYS: No
such file or directory" when running a Git Bash. Thanks to the extra
special `/clangarm64` logic in the Git wrapper, it is merely an
annoyance but does not break anything.

This bug was hidden in the x64 build due to an unrelated change that
fixed the build of the i686 `mingw-w64-git` package in `git-sdk-64`:
https://github.com/git-for-windows/git-sdk-64/commit/c3dc3a1bdcc9. This
commit was done in reaction to the `git-artifacts` build failure at
https://github.com/git-for-windows/git-sdk-64/actions/runs/15236505205/job/42851060773#step:10:207
where the `PATH` variable was not set up in accordance with
`MSYSTEM=MINGW32` and therefore built the x64 variant by mistake when
it should have built the x86 variant. The fix was to ensure that
`/etc/msystem.d/` is included in `build-installers`, which kind of
worked around the problem with the x64 build, but in a fragile way:
Once we stop building the x86 MINGW packages in git-sdk-64, we can drop
all of the x86 bits in git-sdk-64, including `makepkg-git-i686`.

Signed-off-by: Johannes Schindelin <johannes.schindelin@gmx.de>