astextplain: use mimetype to decide helper
We use the file extension to determine which helper to call to convert a
file to text. We currently match an all upper case version and an all
lower case version of each extension, but sometimes variations like
'foo.Rtf' fall though our raster.

To cover all variations in this established way we'd have to add every
permutation of upper and lower case for each extension to the script.
That's a big maintainance burden.
We could also convert the passed filename to a known case and match
against that. Since POSIX has no easy way for sh to do case conversion
in-process we'd have to shell out to tr, awk or similar.

For the antiword helper we already need to make sure the file is of the
type it claims by it's extension. We use file to tell us the mimetype
for this. file can also tell us the types of other files we care about
for astextplain, so we can use it for the decision which helper to use
in the first place.

This fixes https://github.com/git-for-windows/git/issues/5641

Signed-off-by: Matthias Aßhauer <mha1993@live.de>