git-extra: fix the `pkgver` logic again
In 48e2403b3 (git-extra: ignore generated changes for the sake of
`pkgver`, 2025-03-05) I tried to address the fact that the commit that
updates `pkgver` in `PKGBUILD` necessarily touches the `git-extra/`
subdirectory, yet should not be considered for the package version.

What I failed to address is the rare case when `HEAD` is not the same as
`$rev`, i.e. when the last commit that touched the `git-extra/`
directory` is not at the tip of the branch.

Let's accommodate for that, too.

Signed-off-by: Johannes Schindelin <johannes.schindelin@gmx.de>