git-sdk.sh: really accommodate for the recent `pc-msys` -> `pc-cygwin` change
The MSYS2 project switched its triplet from `x86_64-pc-msys` to
`x86_64-pc-cygwin`. We need to accommodate for that.

In 804d5810d (git-sdk.sh: accommodate for the recent `pc-msys` ->
`pc-cygwin` change, 2025-06-22), I already tried to accommodate for
that, but failed to realize that the `chost` triplet is used at two
separate levels of the build directory's absolute path, not only one.

Signed-off-by: Johannes Schindelin <johannes.schindelin@gmx.de>