CI: do test `mingit` when appropriate
In 53038d47a (ci: avoid building the installer multiple times,
2023-08-28), I claimed that the MinGit artifact is already built in the
`sdk-artifacts` matrix.

That is only _kind_ of true: As part of building the `build-installers`
SDK artifact, there _is_ a part where the a minimal SDK is built, but
not MinGit.

Let's build MinGit again when code was touched that changes how it is
built.

Signed-off-by: Johannes Schindelin <johannes.schindelin@gmx.de>