shears: avoid updating other refs
Since Git v2.38.0, support for `rebase.updateRefs` allows enabling the
`--update-refs` option by default. This will update all the local refs
that point to any commits that are rebased in interactive rebases.

However, when using the `shears.sh` script, it is quite common to start
with creating a branch off of the `main` branch via `git switch -c
rebase-to-<version> main` and then running the `shears.sh` script. In
such scenarios, it is undesirable to update the `main` branch, i.e. it
makes a lot of sense to override `rebase.updateRefs`.

Signed-off-by: Johannes Schindelin <johannes.schindelin@gmx.de>