installer: fix the Git for Windows logo that is shown on the upper right
This icon is displayed somehow funny: It's as if the left few pixels are
cut off and then pasted on the right side.

The most likely explanation is that Git for Windows did not use the
advice provided by the InnoSetup documentation at
https://jrsoftware.org/ishelp/topic_setup_wizardsmallimagefile.htm, i.e.
to provide an image that is large enough to accommodate all the
supported Zoom levels.

Let's fix that by providing the `.bmp` files at the recommended
resolutions (and aspect ratios!), and also by allowing InnoSetup to
"stretch" them.

Pointed out by Matthew Cheetham.

Signed-off-by: Johannes Schindelin <johannes.schindelin@gmx.de>