Release Notes: Announce that `git flow` has been dropped
After the "successful" strategy has languished for many years, and after
even a "temporary fork" https://github.com/petervanderdoes/gitflow-avh
(that Git for Windows was forced to switch to) has been archived for
more than two years, even the original repository was archived, as per
https://github.com/nvie/gitflow/issues/3#issuecomment-3400894835, at
_very_ long last.

Vincent Driessen recommends switching to `git-flow-next`, which is
written in Go, and which I am unwilling to distribute with Git for
Windows because 1) it would increase the installer size noticeably, and
2) that strategy seems not to be quite as successful as claimed, given
the low usage numbers.

So let's drop it.

Since there has not been a whole lot of talk about the staleness of `git
flow`, it is safe to assume that there are few users left, if any.
Therefore, I deem a deprecation period unnecessary.

Signed-off-by: Johannes Schindelin <johannes.schindelin@gmx.de>