please.sh: retire the `upgrade` command
We used to upgrade packages using a single command, but that did not
integrate well with GitHub Actions. Therefore we now have the system
where the `monitor-components.yml` workflow monitors for new versions,
opening new tickets if there are, the `/open pr` commands can be used in
those tickets to open Pull Requests in MINGW-packages, MSYS2-packages or
build-extra, and the `/add release-note` command can then update the
release notes.

Therefore we do not need that old shell function anymore.

Signed-off-by: Johannes Schindelin <johannes.schindelin@gmx.de>