sdk cd: use a more generic method to sanitize the `PATH`
When `sdk cd` switches to a project, it tries to adjust the `PATH` to
accommodate whether it is an MSYS or MINGW project (and hence wants to
be compiled using an MSYS or a MINGW compiler).

Naturally, that logic prepends the appropriate `bin` directory to the
`PATH`. To avoid overly long `PATH`s after a couple successive `sdk cd`
invocations, there is logic to avoid listing the same paths multiple
times.

However, that logic is a bit suboptimal due to the limitations of `sed`.
Replace the logic with a complete (and much easier-to-understand) `awk`
construct.

Signed-off-by: Johannes Schindelin <johannes.schindelin@gmx.de>