sdk cd: use the fast MINGW variant of `git` even for MSYS projects
When the `sdk cd` function adjusts the `PATH` so that an MSYS compiler
is used for MSYS projects, an unfortunate side effect is that this also
implies that the MSYS `git.exe` is used in subsequent `git` invocations,
which makes everything quite a bit slower.

Let's prepend the `/cmd` directory to the `PATH` for MSYS projects so
that this is not the case: Even while the MSYS compiler is used,
`git` invocations will use the Git wrapper and hence use the (much
faster) MINGW variant.

Signed-off-by: Johannes Schindelin <johannes.schindelin@gmx.de>