signtool: switch to a still-working time server
It is somewhat surprising that the documentation I found at
https://www.thesslstore.com/knowledgebase/code-signing-time-stamping/sectigo-formerly-comodo-time-stamping-service-expiration-what-to-know-and-what-to-do/
claims that this server stopped working on July 9th, 2019, when we
clearly continued to use it just fine, even earlier today, when Git for
Windows v2.52.0-rc2 was code-signed without any problems.

As per the documentation about the correct URL in
https://www.sectigo.com/resource-library/time-stamping-server?td=sha256
let's use the replacement server.

Signed-off-by: Johannes Schindelin <johannes.schindelin@gmx.de>