make-file-list.sh: use `git-for-windows-addons` package if available
In https://github.com/git-for-windows/MINGW-packages/pull/172, I split
the `mingw-w64-git` package into several parts, preparing for a future
where the same package definition is synchronized between Git for
Windows and MSYS2.

This meant splitting out the Git for Windows-specific parts into a new
package called `mingw-w64-git-for-windows-addons`. This package has the
newly split-out packages as dependencies that are requires to keep
essentially the same files in Git for Windows' installer as before. So
let's use that package instead of `mingw-w64-git`, when available.

The change in this commit includes a minor bug fix in the
`has_pacman_package` function, which did not verify that the directory
it was looking for actually existed.

Signed-off-by: Johannes Schindelin <johannes.schindelin@gmx.de>