Revert "installer: fix rendering of red text and links"
To adapt Git for Windows' installer to InnoSetup v6.6.0, I introduced
some logic that adjusts certain coordinates based on the current font
size. What I failed to realize is that that's exactly the reason why
we're using the `ScaleY()` function.

So let's revert commit 072a15809 (installer: fix rendering of red text
and links, 2025-11-12) in preparation for accommodating InnoSetup v6.6
correctly.

Signed-off-by: Johannes Schindelin <johannes.schindelin@gmx.de>