installer: correctly adjust labels' coordinates to InnoSetup v6.6
With InnoSetup v6.6, the default font size changed. Previously, I tried
to determine that font size dynamically and use it to calculate the
coordinates of the `TLabel` instances reflecting links and red text.

What I failed to realize is that there is no way to query the font size
without the current zoom factor. Therefore, the logic was correct only
at the zoom factor I tested with, but the layout was broken at all other
zoom levels.

So let's go back to hard-coding it, which seems to work for now.

This fixes https://github.com/git-for-windows/git/issues/5961.

Signed-off-by: Johannes Schindelin <johannes.schindelin@gmx.de>