please.sh build-mingw-w64-git: push the tag to both repositories
The `makepkg-mingw` setup is somewhat surprising in that it clones the
Git repository _twice_: Once as a bare repository, next to the
`PKGBUILD`, and then as a non-bare repository (using the bare one as an
alternate) into `src/`.

Historically, we only needed to push the (potentially not-yet-published)
tag to the `src/git/` repository.

However, in an upcoming change, `mingw-w64-git`'s `PKGBUILD` will stop
`SKIP`ping over the checksum of the tag, and verify it first. For this,
the tag needs to be also present in the bare repository, so let's make
it so.

Signed-off-by: Johannes Schindelin <johannes.schindelin@gmx.de>