please.sh build-mingw-w64-git: adjust the checksum in `PKGBUILD`
In preparation for upstreaming `mingw-w64-git` to the MSYS2 project, the
checksum of the tag needs to be recorded in the `PKGBUILD` file instead
of `SKIP`ping over it.

Therefore we need to update said checksum, otherwise the build will fail
before it is started.

Likewise, said checksum now has to be replaced _again_ when building the
source package because the Git repository is replaced by a `.tar.gz` in
that instance.

See https://github.com/git-for-windows/MINGW-packages/pull/172 for the
full details.

Signed-off-by: Johannes Schindelin <johannes.schindelin@gmx.de>