please.sh make-mingw-w64-git: configure `WITH_PDBS`
As part of the effort to upstream `mingw-w64-git` to MSYS2, I changed
the default to _not_ build `.pdb` files by default (because MSYS2 lacks
the `mingw-w64-cv2pdb` package, and it is unlikely that they would
accept it either, given that it requires MSBuild and Visual C to build).

That means that one no longer sets `WITHOUT_PDBS` when skipping
`mingw-w64-pdb`, but `WITH_PDBS` when including it in the build.

For the transition period, ensure that the `mingw-w64-git` package of
Git for Windows is built using the correct value for those flags,
depending whether the DLLs required by cv2pdb are found or not.

Signed-off-by: Johannes Schindelin <johannes.schindelin@gmx.de>