sdk build: set the `WITH_PDBS` flag, too
There is logic to skip building the `.pdb` files of `mingw-w64-git` and
`mingw-w64-curl` in Git for Windows' package definitions.

However, while upstreaming `mingw-w64-git` to the MSYS2 project, it
became obvious that the default should not be to build those `.pdb`
files, and therefore the flag should not be called `WITHOUT_PDBS`, but
instead the default should be to skip them and the flag should be named
`WITH_PDBS`.

Let's handle both flags when `sdk build`ing packages.

Signed-off-by: Johannes Schindelin <johannes.schindelin@gmx.de>